TestTokenizer: test advancement (against infinite loops) - fails yet
